Prove a kernel_wait_progress lemma
Something still seems missing where we don't cover the case where a
thread is in pc.lock_cas but the lock is already held. In that case we
could go to pc.futex_wait and pc.kernel_wait without another thread
acquiring the lock in the meantime. There's nothing in particular saying
that when a thread is woken up the lock is free or some progress has
been made.